chore: refactor required keyword
