+++ lib/vocabularies/validation/required.ts
@@ -1,8 +1,14 @@
 import type {CodeKeywordDefinition, ErrorObject, KeywordErrorDefinition} from "../../types"
 import type {KeywordCxt} from "../../compile/validate"
-import {checkReportMissingProp, reportMissingProp, propertyInData, noPropertyInData} from "../code"
-import {_, str, nil, not, and, or, Name, Code} from "../../compile/codegen"
-import N from "../../compile/names"
+import {
+  checkReportMissingProp,
+  reportMissingProp,
+  propertyInData,
+  noPropertyInData,
+  getSkipCondition,
+  checkMissingProp,
+} from "../code"
+import {_, str, nil, not, Name, Code} from "../../compile/codegen"
 import {checkStrictMode} from "../../compile/util"
 
 export type RequiredError = ErrorObject<
@@ -43,39 +49,12 @@ const def: CodeKeywordDefinition = {
       }
     }
 
-    function getSkipCondition(prop: string): Code | undefined {
-      const propSchema = cxt.parentSchema.properties?.[prop]
-      if (!propSchema) return undefined
-
-      const hasReadOnly = propSchema.readOnly === true
-      const hasWriteOnly = propSchema.writeOnly === true
-
-      if (!hasReadOnly && !hasWriteOnly) return undefined
-
-      const conditions: Code[] = []
-      const oasContext = _`typeof ${N.this} == "object" && ${N.this} && ${N.this}.oas`
-
-      if (hasReadOnly) {
-        conditions.push(_`${oasContext} && ${N.this}.oas.mode === "request"`)
-      }
-
-      if (hasWriteOnly) {
-        conditions.push(_`${oasContext} && ${N.this}.oas.mode === "response"`)
-      }
-
-      if (conditions.length) {
-        return conditions.length === 1 ? conditions[0] : or(...conditions)
-      }
-
-      return undefined
-    }
-
     function allErrorsMode(): void {
       if (useLoop || $data) {
         cxt.block$data(nil, loopAllRequired)
       } else {
         for (const prop of schema) {
-          const skip = getSkipCondition(prop) ?? _`false`
+          const skip = getSkipCondition(cxt, prop) ?? _`false`
           /**
            * Generate a runtime check: validate `required` only when this property
            * should NOT be skipped in the current context (readOnly/writeOnly).
@@ -92,19 +71,7 @@ const def: CodeKeywordDefinition = {
         cxt.block$data(valid, () => loopUntilMissing(missing, valid))
         cxt.ok(valid)
       } else {
-        const checks: Code[] = []
-        for (const prop of schema) {
-          const missingCheck = and(
-            not(getSkipCondition(prop) ?? _`false`),
-            noPropertyInData(gen, data, prop, opts.ownProperties),
-            _`${missing} = ${prop}`
-          )
-
-          checks.push(missingCheck)
-        }
-
-        gen.if(or(...checks))
-
+        gen.if(checkMissingProp(cxt, schema, missing))
         reportMissingProp(cxt, missing)
         gen.else()
       }
